+++ NetSDK/Domain/Split.cs
@@ -9,7 +9,7 @@ namespace NetSDK.Domain
     public class Split: ICloneable
     {
         public string name { get; set; }
-        public long seed { get; set; }
+        public int seed { get; set; }
         public StatusEnum status { get; set; }
         public bool killed { get; set; }
         public string defaultTreatment { get; set; }
